+++ src/common/Icons.jsx
@@ -283,19 +283,19 @@ export function Arrow({width, height, color }) {
     >
       <path
         fillRule="evenodd"
-        clip-rule="evenodd"
+        clipRule="evenodd"
         d="M9.23121 0.278381C9.35228 0.177385 9.49208 0.101239 9.6426 0.0542964C9.79312 0.00735364 9.95141 -0.00946573 10.1084 0.00479903C10.2655 0.0190638 10.4181 0.0641332 10.5577 0.137432C10.6973 0.21073 10.8211 0.31082 10.922 0.431981L14.9216 5.23198C15.027 5.35241 15.1072 5.49269 15.1575 5.64458C15.2079 5.79646 15.2273 5.95688 15.2147 6.1164C15.2022 6.27592 15.1578 6.43131 15.0843 6.57343C15.0108 6.71555 14.9096 6.84154 14.7867 6.94397C14.6637 7.04641 14.5216 7.12323 14.3685 7.16991C14.2155 7.21659 14.0546 7.23219 13.8955 7.21579C13.7363 7.19939 13.582 7.15133 13.4417 7.07442C13.3014 6.99752 13.1779 6.89333 13.0784 6.76798L9.07761 1.96798C8.87401 1.72352 8.77582 1.40822 8.80462 1.09138C8.83342 0.774549 8.98686 0.48212 9.23121 0.278381Z"
         fill={color || "black"}
       />
       <path
         fillRule="evenodd"
-        clip-rule="evenodd"
+        clipRule="evenodd"
         d="M9.23121 11.7216C8.98686 11.5178 8.83342 11.2254 8.80462 10.9086C8.77582 10.5917 8.87401 10.2764 9.07761 10.032L13.0784 5.23196C13.1779 5.10661 13.3014 5.00242 13.4417 4.92552C13.582 4.84861 13.7363 4.80055 13.8955 4.78415C14.0546 4.76775 14.2155 4.78335 14.3685 4.83003C14.5216 4.87671 14.6637 4.95353 14.7867 5.05597C14.9096 5.1584 15.0108 5.28439 15.0843 5.42651C15.1578 5.56863 15.2022 5.72402 15.2147 5.88354C15.2273 6.04305 15.2079 6.20347 15.1575 6.35536C15.1072 6.50724 15.027 6.64753 14.9216 6.76796L10.922 11.568C10.8211 11.6891 10.6973 11.7892 10.5577 11.8625C10.4181 11.9358 10.2655 11.9809 10.1084 11.9951C9.95141 12.0094 9.79312 11.9926 9.6426 11.9456C9.49208 11.8987 9.35228 11.8226 9.23121 11.7216Z"
         fill={color || "black"}
       />
       <path
         fillRule="evenodd"
-        clip-rule="evenodd"
+        clipRule="evenodd"
         d="M12.8 5.99999C12.8 6.31825 12.6736 6.62347 12.4485 6.84852C12.2235 7.07356 11.9182 7.19999 11.6 7.19999L1.99999 7.19999C1.68173 7.19999 1.3765 7.07356 1.15146 6.84851C0.926416 6.62347 0.799987 6.31825 0.799987 5.99999C0.799987 5.68173 0.926416 5.3765 1.15146 5.15146C1.3765 4.92642 1.68173 4.79999 1.99999 4.79999L11.6 4.79999C11.9182 4.79999 12.2235 4.92642 12.4485 5.15146C12.6736 5.3765 12.8 5.68173 12.8 5.99999Z"
         fill={color || "black"}
       />
